Move NeoMathEngine/Platforms.h to NeoML/FineObjLite/FinePlatforms.h
