+++ NeoML/Java/inference/src/main/jni/CMakeLists.txt
@@ -11,5 +11,6 @@ string(TOUPPER ${CMAKE_SYSTEM_NAME} UPPERCASE_CMAKE_SYSTEM_NAME)
 target_compile_definitions(${PROJECT_NAME} PUBLIC _${UPPERCASE_CMAKE_SYSTEM_NAME})
 
 target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../../../../../NeoMathEngine/include)
+target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../../../../../NeoML/include/NeoML/FineObjLite)
 
 target_link_libraries(${PROJECT_NAME} PRIVATE NeoProxy)
