+++ NeoML/include/NeoML/FineObjLite/FinePlatforms.h
