+++ NeoMathEngine/src/CPU/CpuExecutionScope.h
@@ -15,7 +15,6 @@ limitations under the License.
 
 #pragma once
 
-#include <NeoMathEngine/Platforms.h>
 #include <NeoMathEngine/NeoMathEngineDefs.h>
 
 #ifdef NEOML_USE_SSE
