+++ NeoMathEngine/test/src/common/TestParams.cpp
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 --------------------------------------------------------------------------------------------------------------*/
 
-#include <NeoMathEngine/Platforms.h>
+#include <FinePlatforms.h>
 #include "TestParams.h"
 
 namespace NeoMLTest {
